refactor(tauri): decompose root bootstrap module
