+++ apps/desktop/src-tauri/src/commands/agent_sessions.rs
@@ -1,4 +1,5 @@
-use crate::{as_error, run_service_blocking, AppState};
+use crate::app_state::AppState;
+use crate::command_helpers::{as_error, run_service_blocking};
 use host_domain::AgentSessionDocument;
 use std::collections::HashMap;
 use tauri::State;
