+++ apps/desktop/src-tauri/src/commands/build.rs
@@ -1,6 +1,7 @@
-use crate::{
-    as_error, dev_server_emitter, run_service_blocking, AppState, BuildCompletePayload,
-    PullRequestContentPayload, TaskDirectMergePayload,
+use crate::app_state::AppState;
+use crate::command_helpers::{as_error, dev_server_emitter, run_service_blocking};
+use crate::command_payloads::{
+    BuildCompletePayload, PullRequestContentPayload, TaskDirectMergePayload,
 };
 use host_domain::{
     AgentRuntimeKind, AgentSessionStopRequest, BuildSessionBootstrap, DevServerGroupState,
