+++ apps/desktop/src-tauri/src/commands/git/command_handlers.rs
@@ -1,6 +1,7 @@
+use crate::app_state::AppState;
+use crate::command_helpers::run_service_blocking;
 use crate::command_services::error::CommandServiceResult;
 use crate::command_services::git::{self as git_service, requests as git_requests};
-use crate::{run_service_blocking, AppState};
 use tauri::State;
 
 async fn run_git_command<T, F>(operation_name: &'static str, operation: F) -> Result<T, String>
