+++ apps/desktop/src-tauri/src/commands/git/mod.rs
@@ -1,12 +1,4 @@
-mod command_handlers;
-
-pub use command_handlers::{
-    git_abort_conflict, git_commit_all, git_commits_ahead_behind, git_create_worktree,
-    git_fetch_remote, git_get_branches, git_get_current_branch, git_get_diff, git_get_status,
-    git_get_worktree_status, git_get_worktree_status_summary, git_pull_branch, git_push_branch,
-    git_rebase_abort, git_rebase_branch, git_remove_worktree, git_reset_worktree_selection,
-    git_switch_branch,
-};
+pub(crate) mod command_handlers;
 
 #[cfg(test)]
 mod tests;
