+++ apps/desktop/src-tauri/src/commands/git/tests/authorization/cache_invalidation.rs
@@ -3,11 +3,11 @@ use super::super::fixtures::{
     sample_worktree_status_data, seed_authorized_worktree_cache_with_subset,
     setup_command_git_fixture, setup_command_git_fixture_with_mutations, WorktreeStatusResult,
 };
+use crate::command_payloads::{RepoConfigPayload, RepoSettingsPayload};
 use crate::command_services::git::{authorized_worktree_cache, cache_key, resolve_working_dir};
 use crate::commands::workspace::{
     workspace_save_repo_settings, workspace_select, workspace_update_repo_config,
 };
-use crate::{RepoConfigPayload, RepoSettingsPayload};
 use host_domain::GitUpstreamAheadBehind;
 use serde_json::json;
 use std::{fs, path::Path};
