+++ apps/desktop/src-tauri/src/commands/git/tests/fixtures/app.rs
@@ -3,7 +3,7 @@ use std::{
     sync::{Arc, Mutex},
 };
 
-use crate::AppState;
+use crate::app_state::AppState;
 use host_application::AppService;
 use host_domain::TaskStore;
 use host_infra_system::AppConfigStore;
