+++ apps/desktop/src-tauri/src/commands/runtime.rs
@@ -1,4 +1,5 @@
-use crate::{as_error, run_service_blocking, runtime_ensure_failure_kind, AppState};
+use crate::app_state::AppState;
+use crate::command_helpers::{as_error, run_service_blocking, runtime_ensure_failure_kind};
 use host_domain::{
     BeadsCheck, RepoRuntimeHealthCheck, RepoRuntimeStartupStatus, RuntimeCheck, RuntimeDescriptor,
     RuntimeInstanceSummary, SystemCheck, TaskWorktreeSummary,
