+++ apps/desktop/src-tauri/src/commands/system.rs
@@ -1,4 +1,5 @@
-use crate::{as_error, run_service_blocking, AppState};
+use crate::app_state::AppState;
+use crate::command_helpers::{as_error, run_service_blocking};
 use anyhow::{anyhow, Result};
 use host_domain::{SystemOpenInToolId, SystemOpenInToolInfo};
 use serde_json::json;
