+++ apps/desktop/src-tauri/src/headless/task_commands.rs
@@ -4,12 +4,12 @@ use super::command_support::{
     handle_repo_task_reason_operation, request_error, run_command_service_blocking,
     serialize_value, service_error, CommandResult, HeadlessState, RepoPathArgs, RepoTaskArgs,
 };
-use crate::command_services::tasks as task_service;
-use crate::commands::documents::map_plan_subtasks;
-use crate::{
+use crate::command_payloads::{
     BuildCompletePayload, MarkdownPayload, PlanPayload, PullRequestContentPayload,
     TaskDirectMergePayload,
 };
+use crate::command_services::tasks as task_service;
+use crate::commands::documents::map_plan_subtasks;
 use serde::Deserialize;
 use serde_json::{json, Value};
